requirements.txt
# Agent
langchain
langchain-openai
langchain-core
psycopg2
psycopg2-binary
streamlit
python-dotenv
